requirements.txt
python-telegram-bot==20.7
ccxt==3.0.78
sqlalchemy==1.4.46
cryptography==41.0.5